notebooks/pipeline_modules/pipeline.py

"""Main traffic data pipeline."""
import yaml
import json
import logging
from pathlib import Path
from typing import Dict, Any
from datetime import datetime
from dataclasses import asdict
import pandas as pd

from .base import PipelineComponent, PipelineResult
from .validators import DataValidator
from .feature_engineering import FeatureEngineer
from .data_io import DataSaver

class TrafficDataPipeline:
    """Main pipeline orchestrator."""
    
    def __init__(self, config_path: str = 'pipeline_config.yaml'):
        self.config = self._load_config(config_path)
        self.logger = self._setup_logging()
        self.components = self._initialize_components()
        self.execution_history = []
    
    # ... rest of the pipeline implementation ...
