notebooks/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
scikit-learn>=1.1.0
pyarrow>=9.0.0
shap>=0.42.0